+++ CHANGELOG.md
@@ -33,6 +33,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GEOD-1222](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1222)
 - Add pills surrounding on the collapsed Parameters and Sources
   [GEOD-917](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-917)
+- Update the Carousel to include optional Captions and Links
+  [GEOD-1247](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1247)
 
 ### Fixed
 - Ensure the map pans correctly when switching Sensors on the Explore Page